refactor: restructure src directory by feature and add safety to useProjectSession
- Reorganize src/ into feature-based directories:
  - core/ (program, pdas, types)
  - transactions/ (builders)
  - queries/ (balances, eligibility, projects)
  - utils/ (calculations, errors, cache)

- Add safety features to useProjectSession hook:
  - Concurrent fetch prevention with isFetchingRef
  - 30-second timeout protection for slow RPCs
  - Debug logging with [useProjectSession] prefix

- Update all imports across SDK and React hooks
- Update src/index.ts exports for new structure
- All builds pass successfully

This improves code organization, maintainability, and makes useProjectSession
as robust as useLoadedProject.
